DBFS : testé sur un dataguard
Des fonctions ont été déplacées :
    de db/create_pdb.sh vers dblib.sh

Corrections/évolutions.
    db/dbfs/configure_fuse_and_dbfs_mount_point.sh
    db/dbfs/create_crs_resource_for_dbfs.sh
    db/dbfs/create_dbfs.sh
    db/dbfs/oracle_drop_all.sh
    db/dbfs/root_drop_all.sh

+++ db/dbfs/oracle_drop_all.sh
@@ -11,18 +11,25 @@ typeset -r ME=$0
 
 typeset -r str_usage=\
 "Usage : $ME
+\t-db=name
 \t-pdb=name
-\t[-drop_wallet=yes]\tyes|no
+\t[-drop_wallet=no] yes|no
 "
 
+typeset db=undef
 typeset pdb=undef
-typeset drop_wallet=yes
+typeset drop_wallet=no
 
 while [ $# -ne 0 ]
 do
 	case $1 in
 		-emul)
 			EXEC_CMD_ACTION=NOP
+			shift
+			;;
+
+		-db=*)
+			db=$(to_lower ${1##*=})
 			shift
 			;;
 
@@ -55,6 +62,7 @@ must_be_user oracle
 
 exit_if_ORACLE_SID_not_defined
 
+exit_if_param_undef db	"$str_usage"
 exit_if_param_undef pdb	"$str_usage"
 
 exit_if_param_invalid drop_wallet "yes no" "$str_usage"
@@ -69,23 +77,64 @@ fi
 
 . $dbfs_cfg_file
 
+if dataguard_config_available
+then
+	typeset	-r role=$(read_database_role $db)
+else
+	typeset	-r role=primary
+fi
+
 line_separator
 execute_on_all_nodes_v2 -c "fusermount -u /mnt/$pdb"
 LN
 
-sqlplus -s $dbfs_user/$dbfs_password@$service<<EOSQL
-prompt drop filesystem DBFS $dbfs_name
-@?/rdbms/admin/dbfs_drop_filesystem.sql $dbfs_name
-EOSQL
+if [ $role == primary ]
+then
+	sqlplus -s $dbfs_user/$dbfs_password@$service<<-EOSQL
+	prompt drop filesystem DBFS $dbfs_name
+	@?/rdbms/admin/dbfs_drop_filesystem.sql $dbfs_name
+	EOSQL
+	LN
+
+	sqlplus -s sys/Oracle12@$service as sysdba<<-EOSQL
+	prompt drop user $dbfs_user
+	drop user $dbfs_user cascade;
+	prompt drop tbs $dbfs_tbs
+	drop tablespace $dbfs_tbs including contents and datafiles;
+	EOSQL
+	LN
+fi
+
+line_separator
+info "Remove credential for service $service"
+cat<<EOS>/tmp/deletecred.sh
+#!/bin/bash
+echo mkstore -wrl $ORACLE_HOME/oracle/wallet -deleteCredential $service
+mkstore -wrl $ORACLE_HOME/oracle/wallet -deleteCredential $service<<EOP
+$oracle_password
+EOP
+EOS
+chmod u+x /tmp/deletecred.sh
+
+exec_cmd /tmp/deletecred.sh
 LN
 
-sqlplus -s sys/Oracle12@$service as sysdba<<EOSQL
-prompt drop user $dbfs_user
-drop user $dbfs_user cascade;
-prompt drop tbs $dbfs_tbs
-drop tablespace $dbfs_tbs including contents and datafiles;
-EOSQL
-LN
+if [ $gi_count_nodes -ne 1 ]
+then
+	exec_cmd touch $ORACLE_HOME/oracle/wallet/is_cfs
+	ssh ${gi_node_list[0]} test -f $ORACLE_HOME/oracle/wallet/is_cfs
+	if [ $? -ne 0 ]
+	then # ce n'est pas un CFS
+		for node in ${gi_node_list[*]}
+		do
+			exec_cmd scp /tmp/deletecred.sh ${node}:/tmp/deletecred.sh
+			exec_cmd "ssh ${node} '. .bash_profile; /tmp/deletecred.sh'"
+			LN
+		done
+	fi
+	exec_cmd rm $ORACLE_HOME/oracle/wallet/is_cfs
+	LN
+fi
 
 if [ $drop_wallet == yes ]
 then
@@ -98,12 +147,6 @@ then
 	LN
 
 	execute_on_all_nodes_v2 "rm -rf $ORACLE_HOME/oracle/wallet"
-	LN
-else
-	fake_exec_cmd mkstore -wrl $ORACLE_HOME/oracle/wallet -deleteCredential $service
-	mkstore -wrl $ORACLE_HOME/oracle/wallet -deleteCredential $service<<-EOP
-	$oracle_password
-	EOP
 	LN
 fi
